+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+* Added support for generating GIF thumbnails and Image from video files or URL or Assets.
+* Removed Parameter `multiProcess` from `generateGifThumbnail` method.
+* Changed the return type of the `generateGifThumbnail` and `generateImageThumbnail` methods from void to bool, so that the method returns a boolean value indicating whether the thumbnail generation was successful or not.
+
 ## 0.0.3
 
 * Removed `plugin_platform_interface` dependency.